+++ user/data/src/androidTest/kotlin/me/proton/core/user/data/UserManagerImplTests.kt
@@ -79,6 +79,7 @@ class UserManagerImplTests {
 
     private val cryptoContext: CryptoContext = AndroidCryptoContext(
         keyStoreCrypto = object : KeyStoreCrypto {
+            override fun isUsingKeyStore(): Boolean = false
             override fun encrypt(value: String): EncryptedString = value
             override fun decrypt(value: EncryptedString): String = value
             override fun encrypt(value: PlainByteArray): EncryptedByteArray = EncryptedByteArray(value.array.copyOf())